Refactored properties magic into abstract class
#15
